SX: test freshness of generated files
A common issue is that we update `mdn-data` dependency but without regenerating the types. This should improve the situation by adding a test verifying that the generated files are up to date.
