+++ src/sx/package.json
@@ -30,6 +30,7 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "react": "^17.0.1",
-    "react-test-renderer": "^17.0.1"
+    "react-test-renderer": "^17.0.1",
+    "snapshot-diff": "^0.8.1"
   }
 }
